Move cms impl into agents, and test into test/agents because of some fun shenanigans of many tests sharing require-injected modules
